JIMENEZ_ORDONEZ-TEMPLATE-METHOD: Test agregados y README actualizado
